WindowState: fix touchableRegion for full screen apps
When ForceDefaultOrientation is enabled, the rough one-third left side
of a portrait app is untouchable. That's because the touchableRegion
is set to mBounds of TaskStack. From dumpsys window displays we see
the mBounds is set to absolute coordinate like

WINDOW MANAGER DISPLAY CONTENTS (dumpsys window displays)
  Display: mDisplayId=0
    init=1024x768 160dpi cur=590x768 app=590x720 rng=590x518-768x696
    deferred=false layoutNeeded=false

  Application tokens in top down Z order:
    mStackId=1
    mDeferDetach=false
    mFullscreen=true
    mBounds=[217,0][807,768]
  ...

However, Inputflinger calculates the coordinate (x, y) of MotionEntry
in relative value. As a result, events with x < 217 are dropped.

To fix it, don't set touchableRegion to mBounds for full screen apps
as the line 1710 comment of WindowState.java said.

Bug: https://issuetracker.google.com/issues/64817647